+++ packages/invoice-dashboard/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @requestnetwork/invoice-dashboard
 
+## 0.11.10
+
+### Patch Changes
+
+- Fix Lit Encryption and Wallet Connections
+
+## 0.11.9
+
+### Patch Changes
+
+- Fix Lit Protocol Encryption Issues
+
 ## 0.11.8
 
 ### Patch Changes